+++ bindgen/codegen/mod.rs
@@ -3315,6 +3315,8 @@ pub enum EnumVariation {
     Rust {
         /// Indicates whether the generated struct should be `#[non_exhaustive]`
         non_exhaustive: bool,
+        /// Indicates whether the generated struct should be `#[repr(C)]`
+        repr_c: bool,
     },
     /// The code for this enum will use a newtype
     NewType {
@@ -3346,11 +3348,14 @@ impl fmt::Display for EnumVariation {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let s = match self {
             Self::Rust {
-                non_exhaustive: false,
-            } => "rust",
-            Self::Rust {
-                non_exhaustive: true,
-            } => "rust_non_exhaustive",
+                non_exhaustive,
+                repr_c,
+            } => match (*non_exhaustive, *repr_c) {
+                (false, false) => "rust",
+                (false, true) => "rust_repr_c",
+                (true, false) => "rust_non_exhaustive",
+                (true, true) => "rust_non_exhaustive_repr_c",
+            },
             Self::NewType {
                 is_bitfield: true, ..
             } => "bitfield",
@@ -3379,9 +3384,19 @@ impl FromStr for EnumVariation {
         match s {
             "rust" => Ok(EnumVariation::Rust {
                 non_exhaustive: false,
+                repr_c: false,
+            }),
+            "rust_repr_c" => Ok(EnumVariation::Rust {
+                non_exhaustive: false,
+                repr_c: true,
             }),
             "rust_non_exhaustive" => Ok(EnumVariation::Rust {
                 non_exhaustive: true,
+                repr_c: false,
+            }),
+            "rust_non_exhaustive_repr_c" => Ok(EnumVariation::Rust {
+                non_exhaustive: true,
+                repr_c: true,
             }),
             "bitfield" => Ok(EnumVariation::NewType {
                 is_bitfield: true,
@@ -3401,7 +3416,8 @@ impl FromStr for EnumVariation {
                 std::io::ErrorKind::InvalidInput,
                 concat!(
                     "Got an invalid EnumVariation. Accepted values ",
-                    "are 'rust', 'rust_non_exhaustive', 'bitfield', 'consts',",
+                    "are 'rust', 'rust_repr_c', 'rust_non_exhaustive', ",
+                    "'rust_non_exhaustive_repr_c', 'bitfield', 'consts', ",
                     "'moduleconsts', 'newtype' and 'newtype_global'."
                 ),
             )),
@@ -3428,6 +3444,7 @@ struct EnumBuilder {
 enum EnumBuilderKind {
     Rust {
         non_exhaustive: bool,
+        repr_c: bool,
     },
     NewType {
         is_bitfield: bool,
@@ -3473,9 +3490,13 @@ impl EnumBuilder {
                 is_anonymous: enum_is_anonymous,
             },
 
-            EnumVariation::Rust { non_exhaustive } => {
-                EnumBuilderKind::Rust { non_exhaustive }
-            }
+            EnumVariation::Rust {
+                non_exhaustive,
+                repr_c,
+            } => EnumBuilderKind::Rust {
+                non_exhaustive,
+                repr_c,
+            },
 
             EnumVariation::Consts => EnumBuilderKind::Consts {
                 needs_typedef: !has_typedef,
@@ -3686,14 +3707,23 @@ impl EnumBuilder {
 
         // 2. Generate the enum representation
         match self.kind {
-            EnumBuilderKind::Rust { non_exhaustive } => {
+            EnumBuilderKind::Rust {
+                non_exhaustive,
+                repr_c,
+            } => {
                 let non_exhaustive_opt =
                     non_exhaustive.then(attributes::non_exhaustive);
+
+                let repr = if repr_c {
+                    attributes::repr_c()
+                } else {
+                    quote! { #[repr(#enum_repr)] }
+                };
 
                 quote! {
                     // Note: repr is on top of attrs to keep the test expectations diff small.
                     // a future commit could move it further down.
-                    #[repr(#enum_repr)]
+                    #repr
                     #non_exhaustive_opt
                     #( #attrs )*
                     pub enum #enum_ident {
@@ -4096,6 +4126,64 @@ impl CodeGenerator for Enum {
 
         let item = builder.build(ctx, &enum_rust_ty);
         result.push(item);
+
+        if matches!(variation, EnumVariation::Rust { repr_c: true, .. }) {
+            if let Some(layout) = layout {
+                // rustc sizes a fieldless `#[repr(C)]` enum like the target's
+                // default C enum type, i.e. `c_int` on all supported targets.
+                if layout.size != 4 || layout.align != 4 {
+                    warn!(
+                        "enum `{ident}` has size {} and alignment {}, but \
+                         `#[repr(C)]` enums are usually 4 bytes; it and any \
+                         type containing it may get an incompatible layout \
+                         (e.g. due to -fshort-enums or an explicit underlying \
+                         type)",
+                        layout.size, layout.align
+                    );
+                }
+
+                if !ctx.options().layout_tests {
+                    return;
+                }
+
+                let compile_time = ctx.options().rust_features().offset_of;
+                let fn_name = if compile_time {
+                    None
+                } else {
+                    let fn_name = format!("bindgen_test_layout_{ident}");
+                    Some(ctx.rust_ident_raw(fn_name))
+                };
+                let prefix = ctx.trait_prefix();
+                let size_of_expr = quote! {
+                    ::#prefix::mem::size_of::<#ident>()
+                };
+                let align_of_expr = quote! {
+                    ::#prefix::mem::align_of::<#ident>()
+                };
+                let size = layout.size;
+                let align = layout.align;
+                let size_of_err = format!("Size of {ident}");
+                let align_of_err = format!("Alignment of {ident}");
+
+                if compile_time {
+                    result.push(quote! {
+                        #[allow(clippy::unnecessary_operation, clippy::identity_op)]
+                        const _: () = {
+                            [#size_of_err][#size_of_expr - #size];
+                            [#align_of_err][#align_of_expr - #align];
+                        };
+                    });
+                } else {
+                    result.push(quote! {
+                        #[test]
+                        fn #fn_name() {
+                            assert_eq!(#size_of_expr, #size, #size_of_err);
+                            assert_eq!(#align_of_expr, #align, #align_of_err);
+                        }
+                    });
+                }
+            }
+        }
     }
 }