+++ bindgen/lib.rs
@@ -471,7 +471,7 @@ impl Builder {
 
 impl BindgenOptions {
     fn build(&mut self) {
-        const REGEX_SETS_LEN: usize = 29;
+        const REGEX_SETS_LEN: usize = 30;
 
         let regex_sets: [_; REGEX_SETS_LEN] = [
             &mut self.blocklisted_types,
@@ -492,6 +492,7 @@ impl BindgenOptions {
             &mut self.newtype_global_enums,
             &mut self.rustified_enums,
             &mut self.rustified_non_exhaustive_enums,
+            &mut self.rustified_repr_c_enums,
             &mut self.type_alias,
             &mut self.new_type_alias,
             &mut self.new_type_alias_deref,
@@ -529,6 +530,7 @@ impl BindgenOptions {
                     "--rustified-enum-non-exhaustive",
                     "--constified-enum-module",
                     "--constified-enum",
+                    "--rustified-repr-c-enum",
                     "--type-alias",
                     "--new-type-alias",
                     "--new-type-alias-deref",